Move HasGenerateBindingsAttribute to extension method
